Module legion_core::borrow

Atomic runtime borrow checking module. These types implement something akin to RefCell, but are atomically handled allowing them to cross thread boundaries.

Structs

AtomicRefCell
A RefCell implementation which is thread safe. This type performs all the standard runtime borrow checking which would be familiar from using RefCell.

Exclusive

Ref

RefIter

RefIterMut

RefMap

RefMapMut

RefMapMutSet
A set of RefMapMuts

RefMapSet
A set of RefMaps

RefMut

Shared

TryRefIter

TryRefIterMut

Traits

UnsafeClone
Type used for allowing unsafe cloning of internal types