Module legion_core::filter::filter_fns

Functions

any

changed
Creates a filter which includes chunks for which entity data components of type T have changed since the filter was last executed.

component
Creates an entity data filter which includes chunks that contain entity data components of type T.

passthrough

tag
Creates a shared data filter which includes chunks that contain shared data components of type T.

tag_value
Creates a shared data filter which includes chunks that contain specific shared data values.