Struct legion_core::storage::ComponentStorage

pub struct ComponentStorage { /* fields omitted */ }

Stores a chunk of entities and their component data of a specific data layout.

Methods

impl ComponentStorage

pub fn id(&self) -> ChunkId

Gets the unique ID of the chunk.

pub fn len(&self) -> usize

Gets the number of entities stored in the chunk.

pub fn capacity(&self) -> usize

Gets the maximum number of entities that can be stored in the chunk.

pub fn is_full(&self) -> bool

Determines if the chunk is full.

pub fn is_empty(&self) -> bool

Determines if the chunk is empty.

pub fn is_allocated(&self) -> bool

Determines if the internal memory for this chunk has been allocated.

pub fn entities(&self) -> &[Entity]

Gets a slice reference containing the IDs of all entities stored in the chunk.

pub fn components(
&self,
component_type: ComponentTypeId
) -> Option<&ComponentResourceSet>

Gets a component accessor for the specified component type.

pub fn swap_remove(
&mut self,
ComponentIndex: ComponentIndex,
drop: bool
) -> Option<Entity>

Removes an entity from the chunk by swapping it with the last entry.

Returns the ID of the entity which was swapped into the removed entity's position.

pub fn move_entity(
&mut self,
target: &mut ComponentStorage,
index: ComponentIndex
) -> Option<Entity>

Moves an entity from this chunk into a target chunk, moving all compatable components into the target chunk. Any components left over will be dropped.

Returns the ID of the entity which was swapped into the removed entity's position.

pub fn writer(&mut self) -> StorageWriter

Gets mutable references to the internal data of the chunk.