Trait legion_core::world::ComponentLayout

pub trait ComponentLayout: Sized {
    type Filter: for<'a> Filter<ArchetypeFilterData<'a>>;
    fn get_filter(&mut self) -> &mut Self::Filter;

    fn tailor_archetype(&self, archetype: &mut ArchetypeDescription);
}

Describes the types of a set of components attached to an entity.

Associated Types

type Filter: for<'a> Filter<ArchetypeFilterData<'a>>

A filter type which filters archetypes to an exact match with this layout.

Loading content...
Required methods

fn get_filter(&mut self) -> &mut Self::Filter

Gets the archetype filter for this layout.

fn tailor_archetype(&self, archetype: &mut ArchetypeDescription)

Modifies an archetype description to include the components described by this layout.

Loading content...