Trait legion_core::world::IntoComponentSource

pub trait IntoComponentSource {
    type Source: ComponentSource;
    fn into(self) -> Self::Source;
}

An object that can be converted into a ComponentSource.

Associated Types

type Source: ComponentSource

The component source type that can be converted into.

Loading content...
Required methods

fn into(self) -> Self::Source

Converts self into a component source.

Loading content...